requirements.txt
ccxt>=1.40.0
docplex>=2.20.0
web3>=5.23.0
eth-account>=0.5.6
requests>=2.25.1
py-solc-x>=1.1.1
python-dotenv>=0.19.0